pyproject.toml
[project]
name = "bookflightapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.0",
    "uvicorn[standard]>=0.30.0",
    "sqlmodel>=0.0.24",
    "alembic>=1.13.1", # migration cho database
    "bcrypt>=4.1.2", # hash password
    "python-jose[cryptography]>=3.3.0", # JWT
    "pydantic-settings>=2.2.1", # load config từ .env # validate email (nếu có đăng ký user)
    "email-validator>=2.1.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]",
    "python-multipart>=0.0.20",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
